Move download data code into util
